用来参考学习/app 致远期货自动化/双均线回测系统/okx交易所api接口例子/requirements.txt
okx==0.4.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.1 